device/requirements.txt
fastapi
uvicorn[standard]
pydantic
pyyaml
pyserial
